Do not rely on newlines in armor, also make crc optional
We no longer limit armor line to 96 bytes, and also we don't care if we
get really long lines (that are partially returned by ReadLine). Because
of streaming nature of everything there, no big changes were required -
base64 decoding is also streaming, so it doesn't mind if it gets partial
input.