Refactor Signature.FlagsValid... into KeyFlagBits
KeysById will return Key struct that also has KeyFlagBits in there. For
primary keys, flags will be combined from all identity signatures.
Similar things can be done for Subkeys as well, but it is not
implemented as for now, since we do not store all Signatures we read for
Subkeys, only the chosen one.