Load ElGamal Sign or Encrypt keys (#60)
Add ability to read and parse algo 20 keys - ElGamal Encrypt or Sign.
These were deprecated and are not valid anymore. After this change,
these keys are now being loaded and added to BadSubkeys, instead of
failing whole key reading process. If Bad ElGamal is a primary key,
entire load will fail.

This fixes loading key bundles of people who generated such keys pre
~2002 and bad ElGamal key is one of their subkeys. There are no
downsides of loading and verifying such key if it's maintained that it
can't be used for anything.